+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.2
 
 * The first release of sunday_core.
+
+## 0.0.3
+
+* Added streamUsingKey to GetStorage
+
+* edited listen to listenBox